refactor: move _/.config dirs to _/linux and add gpg surpport for dotfiles

+++ scripts/Initialize-Dotfiles.ps1
@@ -6,6 +6,9 @@ param (
   [Parameter()]
   [switch]
   $Uninstall,
+  [Parameter()]
+  [switch]
+  $Gpg,
   [Parameter(Position = 0, ParameterSetName = 'Path')]
   [SupportsWildcards()]
   [string[]]
@@ -19,18 +22,26 @@ param (
 if ($Path) {
   $LiteralPath = Convert-Path $Path
 }
-$root = [System.IO.Path]::GetDirectoryName($PSScriptRoot)
-$root = switch ($true) {
-  $IsWindows { "$root\_\windows"; break }
-  $IsMacOS { "$root/_/macos"; break }
-  default { "$root/_"; break }
-}
+$root = [System.IO.Path]::GetFullPath($PSScriptRoot + '/../_/' + $(switch ($true) {
+      $IsMacOS { 'macos'; break }
+      $IsWindows { 'windows'; break }
+      default { 'linux'; break }
+    }))
 
 if ($Install) {
+  if ($Gpg) {
+    $email = git config --get user.email
+  }
   return $LiteralPath.ForEach{
     $target = $_.Replace($HOME, $root)
     if ($target.Length -eq $_.Length) {
       return Write-Warning "$_ is not a relative path to HOME, skip"
+    }
+    if ($Gpg) {
+      $target += '.gpg'
+      New-Item $target -Force
+      gpg -eo $target -r $email --yes $_
+      return
     }
     # remove existing file or link
     Remove-Item -LiteralPath $target -ea Ignore
@@ -45,16 +56,25 @@ elseif ($Uninstall) {
     if ($target.Length -eq $_.Length) {
       return Write-Warning "$_ is not a relative path to HOME, skip"
     }
+    if ($Gpg) {
+      $target += '.gpg'
+      Remove-Item -LiteralPath $target -Force -ea Ignore
+      return
+    }
     $content = Get-Content -LiteralPath $target -Raw -Force -ea Ignore
     Remove-Item -LiteralPath $_, $target -Force -ea Ignore
     $content > $_
   }
 }
 
-$excludeDirs = ($IsWindows -or $IsMacOS) ? @() : 'shell', 'windows', 'macos'
-fd -HIa -tf -tl $excludeDirs.ForEach{ "-E$_" } --base-directory=$root | ForEach-Object {
+fd -HIa -tf -tl --base-directory=$root | ForEach-Object {
   $file = $_.Replace($root, $HOME)
   $target = (Get-Item -LiteralPath $_ -Force).ResolvedTarget
+  if ($file.EndsWith('.gpg')) {
+    $file = $file.Substring(0, $file.Length - 4)
+    gpg -do $file --yes $target
+    return
+  }
   if ($null -eq $target) {
     throw "Failed to resolve target for $_"
   }
